Add moderncv/ as a copy
I thought about it for a while and ended up using a copy again instead
of a subrepo. Flexibility to overwrite is the key here, so I went this
way.

Copy is from todays (2020-02-11) master branch of the moderncv repo:
    https://github.com/xdanaux/moderncv/

Commit id from the state copied:
    2033517d3f29f9c16af1b3421d52f281e94aa2a5

That represents the last version of moderncv, last updated Nov 16, 2016.