Update the CV itself
I updated and cleaned the tex for the CV itself. Added my last job,
synced between the languages and removed a mistake in the CV itself (was
shifted two months in my emplyoment with PB, not sure if that was
commited but now it is fixed :-)

I'm actually quite ok with it at the moment. The single sections will
still read confusing to anyone who haven't worked in the specific
fields, but I don't want to generalize it more for now.